requirements.txt
google-generativeai>=0.3.0
Pillow>=9.0.0
requests>=2.25.0
ipython>=7.0.0
mcp[cli]